modellingvfm/requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.12
applicationinsights==0.11.9
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
arrow==0.17.0
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
binaryornot==0.4.4
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
clyent==1.2.2
colorama==0.4.3
conda==4.9.2
conda-package-handling==1.7.0
cookiecutter==1.7.2
cryptography @ file:///C:/ci/cryptography_1598892079907/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
Fiona==1.8.13.post1
GDAL==3.0.2
geographiclib==1.50
geopandas @ file:///tmp/build/80754af9/geopandas_1594937497867/work
geopy==2.0.0
graphviz==0.16
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446511143/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598865816163/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jinja2-time==0.2.0
json5==0.9.5
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook @ file:///C:/ci/notebook_1596840175976/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
olefile==0.46
openpyxl==3.0.7
packaging==20.4
pandas @ file:///C:/ci/pandas_1602083338010/work
pandocfilters==1.4.2
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1594298230227/work
pipreqs==0.4.10
poyo==0.5.0
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1598885458782/work
psutil==5.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeps==1.9.13
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks==1.7.1
pyspark==2.4.4
python-dateutil==2.8.1
python-louvain==0.14
python-slugify==4.0.1
pytz==2020.1
pyvis==0.1.8.2
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1598374667791/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
Rtree==0.9.4
ruamel-yaml==0.15.87
Send2Trash==1.5.0
Shapely==1.7.1
sip==4.19.13
six==1.15.0
stdlib-list==0.8.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1600709023549/work
traitlets==4.3.3
-e git+https://NHSXEI@dev.azure.com/NHSXEI/DHSC-SAE-DataSci/_git/TT_data_science@949b4f473c973640c855b09083d6f8a25ea180ca#egg=ttds
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
xmltodict==0.12.0
yarg==0.1.9
zipp==3.1.0